event_manager/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
qrcode>=7.4.2
django-cors-headers>=4.3.1
